Remove duplicate client methods, keep only the async ones
