+++ Transmission.API.RPC/Core/ITransmissionClient.cs
@@ -25,52 +25,53 @@ public interface ITransmissionClient
     /// </summary>
     string Url { get; }
 
+
     /// <summary>
     /// Update blocklist (API: blocklist-update)
     /// </summary>
     /// <returns>Blocklist size</returns>
-    int BlocklistUpdate();
+    Task<int> BlocklistUpdateAsync();
 
     /// <summary>
     /// Close current session (API: session-close)
     /// </summary>
-    void CloseSession();
+    Task CloseSessionAsync();
 
     /// <summary>
     /// Get free space is available in a client-specified folder.
     /// </summary>
     /// <param name="path">The directory to query</param>
-    long FreeSpace(string path);
+    Task<long> FreeSpaceAsync(string path);
 
     /// <summary>
     /// Get information of current session (API: session-get)
     /// </summary>
     /// <returns>Session information</returns>
-    SessionInfo GetSessionInformation();
+    Task<SessionInfo> GetSessionInformationAsync();
 
     /// <summary>
     /// Get session stat
     /// </summary>
     /// <returns>Session stat</returns>
-    Stats GetSessionStatistic();
+    Task<Stats> GetSessionStatisticAsync();
 
     /// <summary>
     /// See if your incoming peer port is accessible from the outside world (API: port-test)
     /// </summary>
     /// <returns>Accessible state</returns>
-    bool PortTest();
+    Task<bool> PortTestAsync();
 
     /// <summary>
     /// Set information to current session (API: session-set)
     /// </summary>
     /// <param name="settings">New session settings</param>
-    void SetSessionSettings(SessionSettings settings);
+    Task SetSessionSettingsAsync(SessionSettings settings);
 
     /// <summary>
     /// Add torrent (API: torrent-add)
     /// </summary>
     /// <returns>Torrent info (ID, Name and HashString)</returns>
-    NewTorrentInfo TorrentAdd(NewTorrent torrent);
+    Task<NewTorrentInfo> TorrentAddAsync(NewTorrent torrent);
 
     /// <summary>
     /// Get fields of torrents from ids (API: torrent-get)
@@ -78,38 +79,38 @@ public interface ITransmissionClient
     /// <param name="fields">Fields of torrents</param>
     /// <param name="ids">IDs of torrents (null or empty for get all torrents)</param>
     /// <returns>Torrents info</returns>
-    TorrentsResult TorrentGet(string[] fields, params int[] ids);
+    Task<TorrentsResult> TorrentGetAsync(string[] fields, params int[] ids);
 
     /// <summary>
     /// Move torrents to bottom in queue  (API: queue-move-bottom)
     /// </summary>
     /// <param name="ids"></param>
-    void TorrentQueueMoveBottom(int[] ids);
+    Task TorrentQueueMoveBottomAsync(int[] ids);
 
     /// <summary>
     /// Move down torrents in queue (API: queue-move-down)
     /// </summary>
     /// <param name="ids"></param>
-    void TorrentQueueMoveDown(int[] ids);
+    Task TorrentQueueMoveDownAsync(int[] ids);
 
     /// <summary>
     /// Move torrents in queue on top (API: queue-move-top)
     /// </summary>
     /// <param name="ids">Torrents id</param>
-    void TorrentQueueMoveTop(int[] ids);
+    Task TorrentQueueMoveTopAsync(int[] ids);
 
     /// <summary>
     /// Move up torrents in queue (API: queue-move-up)
     /// </summary>
     /// <param name="ids"></param>
-    void TorrentQueueMoveUp(int[] ids);
+    Task TorrentQueueMoveUpAsync(int[] ids);
 
     /// <summary>
-    /// Remove torrents (API: torrent-remove)
+    /// Remove torrents
     /// </summary>
     /// <param name="ids">Torrents id</param>
     /// <param name="deleteData">Remove local data</param>
-    void TorrentRemove(int[] ids, bool deleteData = false);
+    Task TorrentRemoveAsync(int[] ids, bool deleteData = false);
 
     /// <summary>
     /// Rename a file or directory in a torrent (API: torrent-rename-path)
@@ -117,13 +118,13 @@ public interface ITransmissionClient
     /// <param name="id">The torrent whose path will be renamed</param>
     /// <param name="path">The path to the file or folder that will be renamed</param>
     /// <param name="name">The file or folder's new name</param>
-    RenameTorrentInfo TorrentRenamePath(int id, string path, string name);
+    Task<RenameTorrentInfo> TorrentRenamePathAsync(int id, string path, string name);
 
     /// <summary>
     /// Set torrent params (API: torrent-set)
     /// </summary>
     /// <param name="settings">Torrent settings</param>
-    void TorrentSet(TorrentSettings settings);
+    Task TorrentSetAsync(TorrentSettings settings);
 
     /// <summary>
     /// Set new location for torrents files (API: torrent-set-location)
@@ -131,49 +132,49 @@ public interface ITransmissionClient
     /// <param name="ids">Torrent ids</param>
     /// <param name="location">The new torrent location</param>
     /// <param name="move">Move from previous location</param>
-    void TorrentSetLocation(int[] ids, string location, bool move);
+    Task TorrentSetLocationAsync(int[] ids, string location, bool move);
 
     /// <summary>
     /// Start recently active torrents (API: torrent-start)
     /// </summary>
-    void TorrentStart();
+    Task TorrentStartAsync();
 
     /// <summary>
     /// Start torrents (API: torrent-start)
     /// </summary>
     /// <param name="ids">A list of torrent id numbers, sha1 hash strings, or both</param>
-    void TorrentStart(object[] ids);
+    Task TorrentStartAsync(object[] ids);
 
     /// <summary>
     /// Start now recently active torrents (API: torrent-start-now)
     /// </summary>
-    void TorrentStartNow();
+    Task TorrentStartNowAsync();
 
     /// <summary>
     /// Start now torrents (API: torrent-start-now)
     /// </summary>
     /// <param name="ids">A list of torrent id numbers, sha1 hash strings, or both</param>
-    void TorrentStartNow(object[] ids);
+    Task TorrentStartNowAsync(object[] ids);
 
     /// <summary>
     /// Stop recently active torrents (API: torrent-stop)
     /// </summary>
-    void TorrentStop();
+    Task TorrentStopAsync();
 
     /// <summary>
     /// Stop torrents (API: torrent-stop)
     /// </summary>
     /// <param name="ids">A list of torrent id numbers, sha1 hash strings, or both</param>
-    void TorrentStop(object[] ids);
+    Task TorrentStopAsync(object[] ids);
 
     /// <summary>
     /// Verify recently active torrents (API: torrent-verify)
     /// </summary>
-    void TorrentVerify();
+    Task TorrentVerifyAsync();
 
     /// <summary>
     /// Verify torrents (API: torrent-verify)
     /// </summary>
     /// <param name="ids">A list of torrent id numbers, sha1 hash strings, or both</param>
-    void TorrentVerify(object[] ids);
+    Task TorrentVerifyAsync(object[] ids);
 }
